requirements.txt
flask
resend
python-dotenv
gunicorn